Updated my compromised address
My wallet was hacked and tokens emptied out with the exception of the D4R NFT. 
I am certain the hacker has a bot that will collect any tokens in my address as soon as it is deposited. 
I would like to replace my compromised one with a new address. 
I also plan on moving my D4R NFT once PR is accepted and merged.